move overmap city functions to new overmap_city.cpp
